requirements.txt
pandas
pyyaml
sqlalchemy
pytz
influxdb